modified the effects of three cards
